Update helpers.php, function a()
In helpers.php is a function a() used for link generation, i. e. by markdown(). In the opening <a> element was a /, which is ignored by many (all? most?) browsers, but leads to interesting result with jQuery inserting of HTML generated by that function.

I stumbled upon this while developing code for a preview of textareas and not being able to show a link to a task (#1).