Update board.js to always keep filter settings
Save filter settings to localStorage everytime filter_apply() is called.

Get and set filters from localStorage when filter_load_events() is called.

This improves the user experience when browsing tasks on a filtered board. No need to set the filters each time the board loads.

Perhaps a checkbox option to initialize this functionality but I think it should be default behavior.