fix invalid HTML when linking to URLs with numeric fragment identifiers
The code tried to convert task numbers to links by applying a regex substitution to the HTML output of Parsedown. This can sometimes break the HTML.

One example is links to URLs with a fragment identifier that starts with a number, such as http://stackoverflow.com/questions/1732348/regex-match-open-tags-except-xhtml-self-contained-tags/1732454#1732454

The fix is to add the additional processing to the markdown renderer itself, so it's only applied to appropriate parts of the text, and never to the generated HTML tags.