Added column description in Task Boards.
Each column in a board may have description which is free form and textual in nature. This field can be used for many purposes like:

-  A general description of that column.
- Entry criteria: When a task qualifies to move to this column.
- Checklist for exiting this column.

This feature adds a column 'column.description' of type text. One can add/edit descriptions to columns from [project]-> Edit Board link.

In the board view, one icon appears next to the column name if a description was added to it. A pop-over on that icon shows the description. Markdown is supported in description however, Write/Preview options are not shown.