Fix flickering and column sizes.
* Avoid flickering in default view mode, by properly setting the default (horizontal-scroll) layout
directly in CSS. This flickering was introduced with #675 pull request. This also fixes the layout
of the public board. Javascript is now used only to activate the 'compact' mode, thus still causing
a slight flickering, but in that mode only.
* Use fixed-layout on the table to improve layout and rendering speed.
* Slightly reduce the width of the swimlane header column, and make it fixed-width.
* Slightly simplify the Javascript/jQuery code for activating compact mode.

This is replaces pull request #688, with only the table layout fix, since I have no proper solution
for the scrolling problem.