better German translation of relations
"ist übergeordnet" is a bit tricky: "Diese Aufgabe ist übergeordnet (1)": is a bad translation as the direction could be in both ways. The correct translation of "is a parent of" would be "ist ein Elternelement von". This seems a bit long, but is the only really correct possibility I can think of. Others, including "ist übergeordnet von" (not correct German IMHO), or "ist ein Kind von" (seems stupid) are not really better. But "ist übergeordnet" is definitely wrong, because in both directions valid, in this context.