Better German
"löschen" means "delete". It is not really clear wht Kanboard does when you click on that link. It deletes the user?? No, it just removes it from that group. Maybe that should be named clearer, even in English. IMHO a simple "remove" (in German "entfernen") would do a better job than "remove this user" - because it is divergent.
So I would either use "remove user from group", or (better) just "remove".
Anyway. The German term should be "entfernen", not "löschen"