Insert date_modification timestamp
When a comment is either created or udpated, we insert the
date_modification timestamp. In addition, we change the sorting to use
the date_modification column by default. This way if the user changes
the sorting in the UI it will show up correctly (based on updated_date
rather than creation date).