Disable auto-assignment in task dialogs
In most cases you don't want to automatically self-assign newly created tasks.
Having to manually change the assignee field every time you want to create a new
task is cumbersome. Additionally, the "assign me" button is kind of useless when
the creator is pre-selected anyway. So the default should be "unspecified".