Increase size of the "users.language" column
The current schema definition for the language column in the users table
is limited to 5 characters, while the longest currently available
language code in Kanboard is 10 characters long - sr_Latn_RS. This
limitation prevents switching to Serbian as the application interface
language.

As per the currently available ISO 639 language codes this column should
be at least 11 characters in size. Examples of language codes of 11
characters in size are en-US-POSIX and shi_Latn_MA.

References:

 - http://www.localeplanet.com/icu/iso639.html