Easier closing of dialogs
Allow closing of dialogs by clicking on the background area around it.

When the dialog is closed in this manner, if it contains a form with changed data present a warning that there are unsaved changes. Likewise if the user attempts to navigate away from the page, eg page reload, bookmark click, address bar entry.

If the dialog does not contain a form or the data remains unchanged no warning messages are shown.