Add the locale en_GB
This adds the locale en_GB where applicable and makes a difference
between GB English and US English.  This might seem like a minor
thing, but it does affect the rendition of things like the calendar
plugin, which has weeks starting on a Sunday for the US locale, but on
a Monday for the GB locale.