Limit assignee drop-down selector scope
On the task details page "Add a sub-task" pop-up modal contains
drop-down (the SELECT HTML element) with "form-user_id" as the `name`
and the `id` attribute value. This element is in conflict with another,
hidden INPUT element with the same `name` and the `id` value. This
causes an issue where clicking the "Me" (Assignee) link / shortcut (A)
element doesn't select your name in the Assignee SELECT element.

Avoiding conflict by changing the SELECT element `id` value is not
possible as the SELECT element is generated using `FormHelper::select()`
function which uses the same value for attribute `name` and `id`.

Given the assignee shortcut feature is used only in two places: edit
task and edit sub-task dialog, the simplest solution is to select the
closest SELECT element to the link / shortcut (A) element or limit the
scope to the currently opened dialog.

This fix uses the latter approach by limiting the scope of targeted
selector.