Advanced bulk task creation (using text) + subtasks.
Allows tags  description and color to be set for each individual task
when using bulk task creation.

The current bulk task creation, create on task per line, and the use the
full line as the task tile.

This commit allows some attributes to be specified in the task line with
the following syntax

    tile [the description] !color :tag1:tag2

It also allows to create subtasks. Each line starting with a `-` will be
added as a subtask of the previous task. A subtask can also contains a
status indicator (similar to orgmode).

Example

    Task1 !blue
    Task2 [Description of task 2]
    Task 3 :tag1:tag2
    - subtask 1
    - [X] already done.
    - [-] in progress.

Create 3  tasks

   Task1 with color set to blue
   Task 2 with the description set to "Description of task 2"
   Task 3 set with the tags, tag1 and tag2.
   Task contains also 3 subtasks, with a different progress status.