Restore PHP 7.3 support
The last update bumped two dependencies leading to newer PHP requirements.
psr/container to 2.0 which requires PHP 7.4 and phpunit to 9.5 requiring
PHP 7.3.

This sets the targeted platform to 7.3 again, so a composer update should
downgrade psr/container and at least restore 7.3 support.

TODO: This is missing a lock file and vendor update.