Fix various action executions failing if the duration parameter isn't set
It's easy to create an action that has nothing set for the duration
by just going "next" in the UI. This results in an empty string being
stored in the DB for the duration, which with PHP 8 fails as it is
multiplied by an integer during the action execution, resulting in:

TypeError: Unsupported operand types: string * int

With older PHP this resulted in a warning and "0" as a result.
By casting to int explicitely we restore that behaviour and we get 0
if the string can't be converted.

Fixes #5454